pyproject.toml
[tool.poetry]
name = "starnavi"
version = "0.1.0"
description = ""
authors = ["cnsr <dmytrenkoyar@gmail.com>"]

[tool.poetry.dependencies]
# this is a hack to make some packages work
python = "^3.8,<3.9"
django-cors-headers = "^3.5.0"
djangorestframework = "^3.11.1"
markdown = "^3.2.2"
django-filter = "^2.3.0"
djangorestframework-simplejwt = "^4.4.0"
djoser = "^2.0.5"
python-dotenv = "^0.14.0"
requests = "^2.24.0"
Faker = "^4.1.3"

[tool.poetry.dev-dependencies]
django-extensions = "^3.0.8"
black = "^20.8b1"
isort = "^5.5.2"
flake8 = "^3.8.3"

[build-system]
requires = ["poetry>=0.12"]
build-backend = "poetry.masonry.api"
